% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzQC.R
\name{fromDatatoMzQC}
\alias{fromDatatoMzQC}
\title{Allow conversion of plain named lists of R objects (from jSON) to mzQC objects}
\usage{
fromDatatoMzQC(mzqc_class, data)
}
\arguments{
\item{mzqc_class}{Prototype of the class to convert 'data' into}

\item{data}{A list of: A datastructure of R lists/arrays as obtained by 'jsonlite::fromJSON()'}
}
\description{
Allow conversion of plain named lists of R objects (from jSON) to mzQC objects
}
\examples{
    data = rmzqc::MzQCcvParameter$new("acc", "myName", "value")
    data_recovered = rmzqc::fromDatatoMzQC(rmzqc::MzQCcvParameter,
                         list(jsonlite::fromJSON(jsonlite::toJSON(data))))

}
