% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obo.R
\name{getCVDictionary}
\alias{getCVDictionary}
\title{Parse the content of 'psi-ms.obo', 'pato.obo', and 'uo.obo' from the 'rmzqc/cv/' folder as ontology and return their union}
\usage{
getCVDictionary()
}
\value{
a data.frame with columns 'id', 'name', 'def', 'parents', 'children' (and many more) which contains the CV entries
}
\description{
See CV_ class to use this function efficiently.
}
