% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obo.R
\name{getCVInfo}
\alias{getCVInfo}
\title{Returns an MzQCcontrolledVocabulary for the currently used CV (see \code{\link{getCVSingleton}})}
\usage{
getCVInfo()
}
\description{
Returns an MzQCcontrolledVocabulary for the currently used CV (see \code{\link{getCVSingleton}})
}
