% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{getCVTemplate}
\alias{getCVTemplate}
\title{Fills a MzQCcvParameter object with id(accession) and name.
The value (if any) needs to be set afterwards.}
\usage{
getCVTemplate(accession, CV = getCVSingleton())
}
\arguments{
\item{accession}{The ID (=accession) of the term in the CV}

\item{CV}{A CV dictionary, as obtained by getCVDictionary(); defaults to the global singleton, which is populated automatically}
}
\value{
An instance of MzQCcvParameter
}
\description{
Fills a MzQCcvParameter object with id(accession) and name.
The value (if any) needs to be set afterwards.
}
