% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obo.R
\name{getDefaultCV}
\alias{getDefaultCV}
\title{Returns an MzQCcontrolledVocabulary for the currently used CV (see \code{\link{getCVSingleton}})}
\usage{
getDefaultCV()
}
\description{
Returns an MzQCcontrolledVocabulary for the currently used CV (see \code{\link{getCVSingleton}})
}
\note{
This function will be deprecated soon. Use \code{\link{getCVInfo}} instead.
}
