% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obo.R
\name{getDefaultCVVersion}
\alias{getDefaultCVVersion}
\title{Obtains the current 'data-version' from the MS-CV shipped with this package}
\usage{
getDefaultCVVersion()
}
\description{
Obtains the current 'data-version' from the MS-CV shipped with this package
}
\examples{
 getDefaultCVVersion() # "4.1.95"

}
