% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obo.R
\name{getLatest_PSICV_URL}
\alias{getLatest_PSICV_URL}
\title{Get the latest PSI-MS CV release URL}
\usage{
getLatest_PSICV_URL()
}
\description{
This may fail (e.g. if no internet connection is available, or URLs became invalid) then 'NULL' will be returned instead of an URL.
A warning may be emitted, if the URL is out of date (i.e. the GitHub API changed).
}
