% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{getOptionalValue}
\alias{getOptionalValue}
\title{Get an optional value from data, using a default if not present}
\usage{
getOptionalValue(data, field_name, default_value = NA)
}
\arguments{
\item{data}{The data structure to extract from}

\item{field_name}{The name of the field to extract}

\item{default_value}{Value to return if the field doesn't exist}
}
\value{
The field value if it exists, otherwise the default_value
}
\description{
Use this function for optional fields where no warning should be generated if missing.
}
\keyword{internal}
