% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{localFileToURI}
\alias{localFileToURI}
\title{Convert a local filename, e.g. "./myData/test.mzML" to a proper URI (e.g. "file:///user/bielow/myData/test.mzML")}
\usage{
localFileToURI(local_filename, must_exist = TRUE)
}
\arguments{
\item{local_filename}{Path to a file (can be relative to current getwd(); or absolute)}

\item{must_exist}{Require the file to exist}
}
\value{
A URI starting with "file:///" followed by an absolute path
}
\description{
Relative filenames are made absolute.
Backslashes as path separators are replaced by forward slashes (as commonly seen on Windows).
}
