% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzQC.R
\name{MzQCanalysisSoftware}
\alias{MzQCanalysisSoftware}
\title{Details of the software used to create the QC metrics}
\description{
Details of the software used to create the QC metrics

Details of the software used to create the QC metrics
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{accession}}{Accession number identifying the term within its controlled vocabulary.}

\item{\code{name}}{Name of the controlled vocabulary term describing the software tool.}

\item{\code{version}}{Version number of the software tool.}

\item{\code{uri}}{Publicly accessible URI of the software tool or documentation.}

\item{\code{description}}{(optional) Definition of the controlled vocabulary term.}

\item{\code{value}}{(optional) Name of the software tool.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MzQCanalysisSoftware-new}{\code{MzQCanalysisSoftware$new()}}
\item \href{#method-MzQCanalysisSoftware-isValid}{\code{MzQCanalysisSoftware$isValid()}}
\item \href{#method-MzQCanalysisSoftware-toJSON}{\code{MzQCanalysisSoftware$toJSON()}}
\item \href{#method-MzQCanalysisSoftware-fromData}{\code{MzQCanalysisSoftware$fromData()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MzQCanalysisSoftware-new"></a>}}
\if{latex}{\out{\hypertarget{method-MzQCanalysisSoftware-new}{}}}
\subsection{Method \code{new()}}{
Constructor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MzQCanalysisSoftware$new(
  accession = NA_character_,
  name = NA_character_,
  version = NA_character_,
  uri = NA_character_,
  description = NA_character_,
  value = NA_character_
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{accession}}{Optional string value for initialization of field accession}

\item{\code{name}}{Optional string value for initialization of field name}

\item{\code{version}}{Optional string value for initialization of field version}

\item{\code{uri}}{Optional string value for initialization of field uri}

\item{\code{description}}{Optional string value for initialization of field description}

\item{\code{value}}{Optional string value for initialization of field value}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MzQCanalysisSoftware-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-MzQCanalysisSoftware-isValid}{}}}
\subsection{Method \code{isValid()}}{
Verifies validity of the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MzQCanalysisSoftware$isValid(context = "MzQCanalysisSoftware")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{context}}{Optional string describing location in mzQC structure that is used for more informative warning texts.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MzQCanalysisSoftware-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-MzQCanalysisSoftware-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Creates JSON file from this object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MzQCanalysisSoftware$toJSON(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Optional parameters for jsonlite:::asJSON()}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MzQCanalysisSoftware-fromData"></a>}}
\if{latex}{\out{\hypertarget{method-MzQCanalysisSoftware-fromData}{}}}
\subsection{Method \code{fromData()}}{
Sets data for this object from plain named lists of R objects
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MzQCanalysisSoftware$fromData(data, context = "MzQCanalysisSoftware")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{A datastructure of R lists/arrays as obtained by 'jsonlite::fromJSON()'}

\item{\code{context}}{Optional string describing location in mzQC structure that is used for more informative warning texts.}
}
\if{html}{\out{</div>}}
}
}
}
