% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzQC.R
\name{MzQCbaseQuality_getMetric}
\alias{MzQCbaseQuality_getMetric}
\title{Extract a certain metric from a runQuality's list of MzQCqualityMetric}
\arguments{
\item{accession}{Accession of the MzQCqualityMetric}

\item{name}{Name of the MzQCqualityMetric (less stable than accession)}
}
\value{
A list of MzQCqualityMetric's which match.
}
\description{
You must provide either the accession or the name of the metric, but not both.
}
\details{
Usually there should be only one MzQCqualityMetric which matches,
however, this function will return all matches.

Note: this function will stop() if not results are found
}
