% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzQC.R
\name{MzQCmetadata}
\alias{MzQCmetadata}
\title{The metadata for a run/setQuality}
\description{
The metadata for a run/setQuality

The metadata for a run/setQuality
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{label}}{Unique name for the run (for runQuality) or set (for setQuality).}

\item{\code{inputFiles}}{Array/list of MzQCinputFile objects}

\item{\code{analysisSoftware}}{Array/list of MzQCanalysisSoftware objects}

\item{\code{cvParameters}}{(optional) Array of cvParameters objects}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MzQCmetadata-new}{\code{MzQCmetadata$new()}}
\item \href{#method-MzQCmetadata-isValid}{\code{MzQCmetadata$isValid()}}
\item \href{#method-MzQCmetadata-toJSON}{\code{MzQCmetadata$toJSON()}}
\item \href{#method-MzQCmetadata-fromData}{\code{MzQCmetadata$fromData()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MzQCmetadata-new"></a>}}
\if{latex}{\out{\hypertarget{method-MzQCmetadata-new}{}}}
\subsection{Method \code{new()}}{
Constructor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MzQCmetadata$new(
  label = NA_character_,
  inputFiles = list(),
  analysisSoftware = list(),
  cvParameters = list()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{label}}{Optional string value for initialization of field label}

\item{\code{inputFiles}}{Optional list value for initialization of field inputFiles}

\item{\code{analysisSoftware}}{Optional list value for initialization of field analysisSoftware}

\item{\code{cvParameters}}{Optional list value for initialization of field cvParameters}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MzQCmetadata-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-MzQCmetadata-isValid}{}}}
\subsection{Method \code{isValid()}}{
Verifies validity of the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MzQCmetadata$isValid(context = "MzQCmetadata")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{context}}{Optional string describing location in mzQC structure that is used for more informative warning texts.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MzQCmetadata-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-MzQCmetadata-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Creates JSON file from this object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MzQCmetadata$toJSON(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Optional parameters for jsonlite:::asJSON()}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MzQCmetadata-fromData"></a>}}
\if{latex}{\out{\hypertarget{method-MzQCmetadata-fromData}{}}}
\subsection{Method \code{fromData()}}{
Sets data for this object from plain named lists of R objects
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MzQCmetadata$fromData(data, context = "MzQCmetadata")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{A datastructure of R lists/arrays as obtained by 'jsonlite::fromJSON()'}

\item{\code{context}}{Optional string describing location in mzQC structure that is used for more informative warning texts.}
}
\if{html}{\out{</div>}}
}
}
}
