% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{readMZQC}
\alias{readMZQC}
\title{Read a JSON file in mzQC format into an MzQCmzQC root object}
\usage{
readMZQC(filepath)
}
\arguments{
\item{filepath}{A filename (with path) to write to.}
}
\value{
An MzQCmzQC root object from which all the data can be extracted/manipulated
}
\description{
Read a JSON file in mzQC format into an MzQCmzQC root object
}
