% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{readMZQCFromJSON}
\alias{readMZQCFromJSON}
\title{Read a JSON object in mzQC format into an MzQCmzQC root object}
\usage{
readMZQCFromJSON(json_obj)
}
\arguments{
\item{json_obj}{A generic R object (list of lists...)}
}
\value{
An MzQCmzQC root object from which all the data can be extracted/manipulated
}
\description{
Read a JSON object in mzQC format into an MzQCmzQC root object
}
