% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{removeFileSuffix}
\alias{removeFileSuffix}
\title{Removes the last suffix (including the last dot) from a filename.
If no dot exists, the full string is returned.}
\usage{
removeFileSuffix(filepath)
}
\arguments{
\item{filepath}{A filename (with optional path -- which is retained)}
}
\value{
The input with removed suffix
}
\description{
Removes the last suffix (including the last dot) from a filename.
If no dot exists, the full string is returned.
}
\examples{
 removeFileSuffix("test.tar.gz")  # --> 'test.tar'
 removeFileSuffix("test.mzML")  # --> 'test'
 removeFileSuffix("/path/to/test.mzML")  # --> '/path/to/test'
 removeFileSuffix("test_no_dot")  # --> 'test_no_dot'

}
