% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{removeIfExists}
\alias{removeIfExists}
\title{Remove a file, if it exists (useful for temporary files which may or may not have been created)}
\usage{
removeIfExists(tmp_filename)
}
\arguments{
\item{tmp_filename}{A path to a local file}
}
\value{
NULL if file is missing, otherwise TRUE/FALSE depening on successful removal
}
\description{
Remove a file, if it exists (useful for temporary files which may or may not have been created)
}
