% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmzqc-package.R
\docType{package}
\name{rmzqc-package}
\alias{rmzqc}
\alias{rmzqc-package}
\title{rmzqc: Creation, Reading and Validation of 'mzqc' Files}
\description{
Reads, writes and validates 'mzQC' files. The 'mzQC' format is a standardized file format for the exchange, transmission, and archiving of quality metrics derived from biological mass spectrometry data, as defined by the HUPO-PSI (Human Proteome Organisation - Proteomics Standards Initiative) Quality Control working group. See \url{https://hupo-psi.github.io/mzQC/} for details.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MS-Quality-hub/rmzqc}
  \item Report bugs at \url{https://github.com/MS-Quality-hub/rmzqc/issues}
}

}
\author{
\strong{Maintainer}: Chris Bielow \email{chris.bielow@bsc.fu-berlin.de} (\href{https://orcid.org/0000-0001-5756-3988}{ORCID})

Other contributors:
\itemize{
  \item David Jimenez-Morales \email{biodavidjm@gmail.com} (\href{https://orcid.org/0000-0003-4356-6461}{ORCID}) [reviewer, contributor]
}

}
\keyword{internal}
