% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmzqc.R
\docType{package}
\name{rmzqc}
\alias{rmzqc}
\title{rmzqc: A package for reading, validating, and writing mzQC files.}
\description{
The core function of the package is reading mzQC files into an RefClasses wrapped data structure and
writing such data to file again.
}
