% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{toAnalysisSoftware}
\alias{toAnalysisSoftware}
\title{From an ID, e.g. "MS:1003162" (for PTX-QC), and some additional information,
create an 'analysisSoftware' node for mzQC}
\usage{
toAnalysisSoftware(id, version = "unknown", uri = NULL, value = NA_character_)
}
\arguments{
\item{id}{The CV accession}

\item{version}{The version of the tool which created the metric/mzQC}

\item{uri}{URI to the homepage, or if NULL (default), will be extracted from the definition in the PSI MS-CV (if possible)}

\item{value}{An optional name for the software (if different from the CV's name)}
}
\value{
An MzQCanalysisSoftware object
}
\description{
From an ID, e.g. "MS:1003162" (for PTX-QC), and some additional information,
create an 'analysisSoftware' node for mzQC
}
\examples{
 # use 'version = packageVersion("PTXQC")' if the package is installed
 toAnalysisSoftware(id = "MS:1003162", version = "1.0.12")

}
