% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator.R
\name{validateFromFile}
\alias{validateFromFile}
\title{Syntactically validates an mzQC document which is present as a file.}
\usage{
validateFromFile(filepath, verbose = TRUE)
}
\arguments{
\item{filepath}{A path to a file (e.g. "c:/my.mzQC", or "test.mzQC")}

\item{verbose}{Show extra information if validation fails}
}
\value{
TRUE/FALSE if validation was successful/failed
}
\description{
The returned TRUE/FALSE has additional attributes in case of errors.
Use attributes(result) to access them.
}
