% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator.R
\name{validateFromObj}
\alias{validateFromObj}
\title{Syntactically validates an mzQC document which is already in memory as mzQC root object, as obtained by, e.g. readMZQC().}
\usage{
validateFromObj(mzqc_root, verbose = TRUE)
}
\arguments{
\item{mzqc_root}{An mzQC root object}

\item{verbose}{Show extra information if validation fails}
}
\value{
TRUE/FALSE if validation was successful/failed
}
\description{
This method is less performant than validateFromString,
because it needs to convert the R object to
a JSON string first.
}
\details{
The returned TRUE/FALSE has additional attributes in case of errors.
Use attributes(result) to access them.
}
