% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator.R
\name{validateFromString}
\alias{validateFromString}
\title{Syntactically validates an mzQC document which is already in memory as JSON string.
e.g. the string "\{ mzQC : \{\}\}"}
\usage{
validateFromString(JSON_string, verbose = TRUE)
}
\arguments{
\item{JSON_string}{A string which contains JSON (multiple lines allowed)}

\item{verbose}{Show extra information if validation fails}
}
\value{
TRUE/FALSE if validation was successful/failed
}
\description{
If the string object passed into this function contains multiple elements (length > 1).
then they will be concatenated using '\\n' before validation.
}
\details{
The returned TRUE/FALSE has additional attributes in case of errors.
Use attributes(result) to access them.
}
