% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{writeMZQC}
\alias{writeMZQC}
\title{Writes a full mzQC object to disk.}
\usage{
writeMZQC(filepath, mzqc_obj)
}
\arguments{
\item{filepath}{A filename (with path) to write to.}

\item{mzqc_obj}{An mzQC object, which is serialized to JSON and then written to disk}
}
\description{
The filename should have an '.mzQC' as suffix (warning otherwise).
}
