% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzQC.R
\name{fromDatatoMzQCobj}
\alias{fromDatatoMzQCobj}
\title{Allow conversion of a plain R object (obtained from jSON) to an mzQC object}
\usage{
fromDatatoMzQCobj(mzqc_class, data)
}
\arguments{
\item{mzqc_class}{Prototype of the class to convert 'data' into}

\item{data}{A datastructure of R lists/arrays as obtained by 'jsonlite::fromJSON()'}
}
\description{
If you have a list of elements, call fromDatatoMzQC.
}
\examples{
 data = MzQCcvParameter$new("acc", "myName", "value")
 data_recovered = fromDatatoMzQCobj(MzQCcvParameter, jsonlite::fromJSON(jsonlite::toJSON(data)))
 data_recovered

}
