% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obo.R
\name{getCVInfo}
\alias{getCVInfo}
\title{Returns an \code{\link{MzQCcontrolledVocabulary}} for the currently used CV (see \code{\link{getCVSingleton}})
using \code{getCVSingleton()$getData()$URI} and \verb{$version}.}
\usage{
getCVInfo()
}
\description{
Returns an \code{\link{MzQCcontrolledVocabulary}} for the currently used CV (see \code{\link{getCVSingleton}})
using \code{getCVSingleton()$getData()$URI} and \verb{$version}.
}
