% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obo.R
\name{getDefaultCV}
\alias{getDefaultCV}
\title{Returns an MzQCcontrolledVocabulary which points to the PSI-MS CV which is currently shipped with this package}
\usage{
getDefaultCV()
}
\description{
Returns an MzQCcontrolledVocabulary which points to the PSI-MS CV which is currently shipped with this package
}
