% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{getQualityMetricTemplate}
\alias{getQualityMetricTemplate}
\title{Fills a MzQCqualityMetric object with id(accession) and name.
The value (if any) and unit (if any) need to be set afterwards.}
\usage{
getQualityMetricTemplate(accession, CV = getCVSingleton())
}
\arguments{
\item{accession}{The ID (=accession) of the term in the CV}

\item{CV}{A CV dictionary, as obtained by getCVDictionary(); defaults to the global singleton, which is populated automatically}
}
\value{
An instance of MzQCqualityMetric
}
\description{
Fills a MzQCqualityMetric object with id(accession) and name.
The value (if any) and unit (if any) need to be set afterwards.
}
