% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzQC.R
\docType{class}
\name{MzQCrunQuality-class}
\alias{MzQCrunQuality-class}
\alias{MzQCrunQuality}
\title{The runQuality object. Use to report metrics for individual runs which are independent of other runs.}
\description{
The runQuality object. Use to report metrics for individual runs which are independent of other runs.
}

