% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzQC.R
\name{MzQCrunQuality}
\alias{MzQCrunQuality}
\title{A runQuality object. Use to report metrics for individual runs which are independent of other runs.}
\description{
The object is an alias for MzQCbaseQuality.
}
\section{Super class}{
\code{\link[rmzqc:MzQCbaseQuality]{rmzqc::MzQCbaseQuality}} -> \code{MzQCrunQuality}
}
