% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzQC.R
\docType{class}
\name{MzQCsetQuality-class}
\alias{MzQCsetQuality-class}
\alias{MzQCsetQuality}
\title{The setQuality object. Use it for metrics which are specific to sets, i.e. only for values which
only make sense in the set context and cannot be stored as runQuality (see mzQC spec doc).}
\description{
The setQuality object. Use it for metrics which are specific to sets, i.e. only for values which
only make sense in the set context and cannot be stored as runQuality (see mzQC spec doc).
}

