% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzQC.R
\name{MzQCsetQuality}
\alias{MzQCsetQuality}
\title{A setQuality object. Use it for metrics which are specific to sets, i.e. only for values which
only make sense in the set context and cannot be stored as runQuality (see mzQC spec doc).}
\description{
The object is an alias for MzQCbaseQuality.
}
\section{Super class}{
\code{\link[rmzqc:MzQCbaseQuality]{rmzqc::MzQCbaseQuality}} -> \code{MzQCsetQuality}
}
