% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{writeMZQC}
\alias{writeMZQC}
\title{Writes a full mzQC object to disk.}
\usage{
writeMZQC(filepath, mzqc_obj)
}
\arguments{
\item{filepath}{A filename (with optional path) to write to.}

\item{mzqc_obj}{An MzQCmzQC root object, which is serialized to JSON and then written to disk}
}
\description{
You can in theory also provide any mzQC subelement,
but the resulting mzQC file will not validate since its incomplete.
}
\details{
The filename should have '.mzQC' (case sensitive) as suffix. There will be a warning otherwise.
}
